tests: a cache hit leaves no reference the region graph cannot explain
`teardown_leaves_no_unexplained_references` compiles the stdlib so its verdict
does not move with the state of a cache file, and that is what stops it seeing
the reload's own claim. A hit rebuilds the stdlib's closures, templates and
capture cells through an allocation context of its own, and the Rust code that
keeps them releases nothing, so the region survives the sweep at rc=1 with no
in-edge at all.

The residue count cannot see it either: one region on a residue of over a
hundred, indistinguishable from the reference cycle that explains the rest.

The run/root/teardown/read sequence both gates share is now one helper, so the
cache-hit gate asks the compiled gate's question rather than a paraphrase of it.
